Generate openstack cloud config in correct INI format
The current cloud config generated by RKE is invalid for openstack because it marshals the config key/value into a json file where the openstack provider expects a INI from the gcfg.

This correctly generates the correct text file format when provider name is "openstack" and fallbacks to the former generic config format otherwise.

This factors out the current code for generic cloud config and add test units.